feat(firehose/logs): message searching
